demo/streamlit/requirements.txt
transformers
torch
praw
python-dotenv
streamlit=1.19.0
pandas
